+++ dashboard/src/components/presentational/CopyToClipboardButton/index.ts
@@ -1,0 +1,1 @@
+export { default as CopyToClipboardButton } from './CopyToClipboardButton';